gmte_binary

Performs analyses for the Triangulation WIthin A STudy (TWIST) framework to calcuate the ‘genetically moderated treatment effect’ (GMTE) for a binary outcome. The gmte_binary function returns an object of class twistR_GMTE, containing effect estimates from the individual tests (such as RGMTE) and the results when combinations are performed (such as RGMTE+MR).

Usage

gmte_binary(
  Y,
  T,
  G,
  Z,
  D,
  Link = "logit",
  alpha = 0.05,
  doCAT = FALSE,
  verbose = FALSE
)

Arguments

Y

The binary outcome variable name (string) which appears in data.frame D.

T

The treatment variable name (string) which appears in data.frame D. Assumed to be binary.

G

The genotype variable name (string) which appears in data.frame D. Normally binary (e.g. comparing homozygous rare individuals to the rest of the population) but can be additive (0, 1, 2) or a score of multiple variants if desired.

Z

A string containing the model covariates to appear in the glm() models (for example "age+sex"). All need to be in data.frame D.

D

A data.frame containing the above variables.

Link

Link function for the glm() - needs to be one of "logit","probit" or "identity". If unspecified the default is "logit".

alpha

The p-value threshold for the chi-square test, estimating whether two estimates should be combined. Default is 0.05.

doCAT

Run the CAT analysis? Default is FALSE.

verbose

Return lots of output - useful for error checking. Default is FALSE.

Value

An object of class twistR_GMTE containing the following components:

CAT

The summary statistics from the Corrected As Treated (CAT) analysis.

GMTE1

The summary statistics from the GMTE(1) analysis (i.e. in the treated individuals).

GMTE0

The summary statistics from the GMTE(0) analysis (i.e. in the untreated individuals).

MR

The summary statistics from the MR analysis.

RGMTE

The summary statistics from the Robust GMTE analysis (GMTE1 corrected for GMTE0).

FullCombined

The combined summary statistics from all analyses performed, inclduing combinations.

References

Bowden, J., et al., The Triangulation WIthin A STudy (TWIST) framework for causal inference within Pharmacogenetic research. PLoS Genetics. https://doi.org/10.1371/journal.pgen.1009783

Author

Jack Bowden; Luke Pilling.

Examples

# Example using a binary outcome (high LDL), binary treatment (statins), and binary genotype (SLCO1B1*5 homozygotes) variables
Y="ldl_high"
T="statin"
G="slco1b1_5_hmz"
Z="age+PC1+PC2+PC3+PC4+PC5+PC6+PC7+PC8+PC9+PC10"
Link="logit"
results=gmte_binary(Y,T,G,Z,D,Link)
#> TWIST (Triangulation WIthin A STudy) analysis in R - binary outcome
#> Error in gmte_binary(Y, T, G, Z, D, Link): D needs to be a data.frame